# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger API documentation for EchoCompany",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
        "/articles/?page=": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Выводит список всех статей",
                "operationId": "articlesList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы пагинации",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Получение определенной статьи по ее id",
                "operationId": "articleId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Номер страницы пагинации",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/articles/search/{field}/{value}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Поиск необходимой статьи по одному из четырех параметров",
                "operationId": "articlesList",
                "parameters": [
                    {
                        "name": "field",
                        "in": "path",
                        "description": "Допустимые поля:author,category,title,slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "value",
                        "in": "path",
                        "description": "Значение для поиска",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Article is not found"
                    }
                }
            }
        },
        "/articles/sort/{value}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Сортировка статей по одному из 4 параметров. В Swagger сортировка не отображается, но если отправить запрос напрямую к API, то вернется отсортированный массив",
                "description": "Сортировка статей по одному из 4 параметров",
                "operationId": "baa1b08c2e2d12ff4bd2945dbe2e52ec",
                "parameters": [
                    {
                        "name": "value",
                        "in": "path",
                        "description": "Допустимые значения для сортировки:id,-id,title,-title. Знак минус означает обратную сортировку",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Article is not found"
                    }
                }
            }
        },
        "/authors/?page=": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Выводит список всех авторов",
                "operationId": "authorsList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы пагинации",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/authors/{author_id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Получение определенного автора по его id",
                "description": "Получение определенного автора по его id",
                "operationId": "138288da91eac62e7d0ea55295079a4f",
                "parameters": [
                    {
                        "name": "author_id",
                        "in": "path",
                        "description": "Author id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Author is not found"
                    }
                }
            }
        },
        "/authors/search/{field}/{value}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Получение необходимого автора по одному из двух параметров",
                "description": "Получение необходимого автора по одному из двух параметров",
                "operationId": "420e16f7c726b9a2bacb1ecc0501cc40",
                "parameters": [
                    {
                        "name": "field",
                        "in": "path",
                        "description": "Допустимые поля:name,slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "value",
                        "in": "path",
                        "description": "Значение для поиска",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Article is not found"
                    }
                }
            }
        },
        "/authors/sort/{value}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Сортировка авторов по одному из 4 параметров. В Swagger сортировка не отображается, но если отправить запрос напрямую к API, то вернется отсортированный массив",
                "description": "Сортировка авторов по одному из 4 параметров",
                "operationId": "36112d705c81b94d03f24f818d780341",
                "parameters": [
                    {
                        "name": "value",
                        "in": "path",
                        "description": "Допустимые значения для сортировки:id,-id,name,-name.Знак минус означает обратную сортировку",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Author is not found"
                    }
                }
            }
        },
        "/categories/?page=": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Выводит список всех категорий",
                "operationId": "categoriesList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы пагинации",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/categories/{category_id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Получение определенной категории по ее id",
                "description": "Получение определенной категории по ее id",
                "operationId": "4d95bac56cddb18d477341d2ad48ef88",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Category is not found"
                    }
                }
            }
        },
        "/categories/search/{field}/{value}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Поиск необходимой категории по одному из двух параметров",
                "description": "Поиск необходимой категории по одному из двух параметров",
                "operationId": "4b5996a6f69fedfed2dce1013cd618ca",
                "parameters": [
                    {
                        "name": "field",
                        "in": "path",
                        "description": "Допустимые поля:title,slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "value",
                        "in": "path",
                        "description": "Значение для поиска",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Category is not found"
                    }
                }
            }
        },
        "/categories/sort/{value}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Сортировка категорий по одному из 4 параметров. В Swagger сортировка не отображается, но если отправить запрос напрямую к API, то вернется отсортированный массив",
                "description": "Сортировка категорий по одному из 4 параметров",
                "operationId": "2ebf3cb0f36430096a53c7557b7c684d",
                "parameters": [
                    {
                        "name": "value",
                        "in": "path",
                        "description": "Допустимые значения для сортировки:id,-id,title,-title. Знак минус означает обратную сортировку",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Article is not found"
                    }
                }
            }
        },
        "/": {
            "get": {
                "description": "Home page",
                "responses": {
                    "default": {
                        "description": "Welcome page"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "X-APP-ID": {
                "type": "apiKey",
                "name": "X-APP-ID",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Authors",
            "description": "Поиск авторов"
        },
        {
            "name": "Articles",
            "description": "Поиск статей"
        },
        {
            "name": "Categories",
            "description": "Поиск категорий"
        }
    ],
    "security": [
        []
    ]
}